+++ src/main/java/com/vanym/paniclecraft/client/ClientProxy.java
@@ -1,9 +1,12 @@
 package com.vanym.paniclecraft.client;
 
+import java.util.Optional;
+
 import com.vanym.paniclecraft.Core;
 import com.vanym.paniclecraft.client.command.ClientCommandMod3;
-import com.vanym.paniclecraft.client.command.ClientCommandVersion;
+import com.vanym.paniclecraft.command.CommandDev;
 import com.vanym.paniclecraft.command.CommandMod3;
+import com.vanym.paniclecraft.command.CommandVersion;
 import com.vanym.paniclecraft.core.CommonProxy;
 import com.vanym.paniclecraft.core.ModConfig;
 import com.vanym.paniclecraft.core.component.IModComponent;
@@ -16,11 +19,16 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 public class ClientProxy extends CommonProxy {
     
     public ClientCommandMod3 command;
+    public Optional<CommandDev> devCommand = Optional.empty();
     
     @Override
     public void preInit(ModConfig config) {
         this.command = new ClientCommandMod3();
-        this.command.addSubCommand(new ClientCommandVersion());
+        this.command.addSubCommand(new CommandVersion());
+        if (Core.instance.devCommand.isPresent()) {
+            this.devCommand = Optional.of(new CommandDev());
+            this.command.addSubCommand(this.devCommand.get());
+        }
         ClientCommandHandler.instance.registerCommand(this.command);
         for (IModComponent component : Core.instance.getComponents()) {
             component.preInitClient(config);
@@ -47,5 +55,10 @@ public class ClientProxy extends CommonProxy {
     @Override
     public CommandMod3 getCommand() {
         return this.command;
+    }
+    
+    @Override
+    public Optional<CommandDev> getDevCommand() {
+        return this.devCommand;
     }
 }