+++ src/main/java/com/vanym/paniclecraft/command/CommandDev.java
@@ -1,0 +1,10 @@
+package com.vanym.paniclecraft.command;
+
+public class CommandDev extends CommandMod3 {
+    
+    public static final String NAME = "dev";
+    
+    public CommandDev() {
+        super(NAME);
+    }
+}