+++ src/main/java/com/vanym/paniclecraft/command/CommandVersion.java
@@ -23,6 +23,11 @@ public class CommandVersion extends CommandBase {
     }
     
     @Override
+    public boolean checkPermission(MinecraftServer server, ICommandSender sender) {
+        return true;
+    }
+    
+    @Override
     public void execute(MinecraftServer server, ICommandSender sender, String[] args)
             throws CommandException {
         if (args.length > 0) {