+++ src/main/java/com/vanym/paniclecraft/server/ServerProxy.java
@@ -1,6 +1,9 @@
 package com.vanym.paniclecraft.server;
 
+import java.util.Optional;
+
 import com.vanym.paniclecraft.Core;
+import com.vanym.paniclecraft.command.CommandDev;
 import com.vanym.paniclecraft.command.CommandMod3;
 import com.vanym.paniclecraft.core.CommonProxy;
 
@@ -13,5 +16,10 @@ public class ServerProxy extends CommonProxy {
     @Override
     public CommandMod3 getCommand() {
         return Core.instance.command;
+    }
+    
+    @Override
+    public Optional<CommandDev> getDevCommand() {
+        return Core.instance.devCommand;
     }
 }